Added sixtrack assimilator and its Makefile
the current version is the one which started operation on 2016-08-18, at ~20:00.
The makefile has been changed according to the modifications found while compiling.

All previous versions of the assimilator have not been deleted but
are stored in before_git_repo, for future, quick lookups. This directory
contains only versions of the assimilator older than the present one, and
nothing more.

Whenever generating a new version of the sixtrack assimilator, please use all the
capabilities of git (e.g. do not add the current version to the before_git_repo dir).