bug fix in run_six.sh: re-submission to lsf after boinc (thanks for FVDVeken)
The bug was that, even if the -S option was requested by the user (ie
resubmit only those jobs which did not generate any output so far),
the existing results were deleted. This was due to the fact that
the submitCreateRundir() was wrongly called, ie checking only that
no input files existed, instead of checking also that output files
exist even if no (lsf) input is generated.

At the same time, improved a bit the logics of checks.